src/main/python/airflow/tasks/youtubefetch/requirements.txt
google-cloud-secret-manager>=2.7.0
google-cloud-storage>=1.42.0
google-auth==2.26.2
